+++ src/main/java/urlshortener/web/SystemInformationController.java
@@ -1,4 +1,4 @@
-package urlshortener.information;
+package urlshortener.web;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,6 +9,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+import urlshortener.domain.Information;
 import urlshortener.service.ClickService;
 import urlshortener.service.ShortURLService;
 
@@ -16,7 +17,7 @@ import urlshortener.service.ShortURLService;
 @Configuration
 @EnableScheduling
 @Component
-public class SystemInformation {
+public class SystemInformationController {
 
   private final AtomicInteger numClicks;
   private final AtomicInteger numUsers;
@@ -25,7 +26,7 @@ public class SystemInformation {
   private final ClickService clickService;
   private final ShortURLService shortUrlService;
 
-  public SystemInformation(ClickService clickService, ShortURLService shortUrlService) {
+  public SystemInformationController(ClickService clickService, ShortURLService shortUrlService) {
     this.numClicks = new AtomicInteger(0);
     this.numURLs = new AtomicInteger(0);
     this.numUsers = new AtomicInteger(0);
@@ -52,39 +53,4 @@ public class SystemInformation {
     return list;
   }
 
-  public class Information {
-    public String name;
-    public String description;
-    public AtomicInteger number;
-
-    public Information(String name, String description, AtomicInteger number) {
-      this.name = name;
-      this.description = description;
-      this.number = number;
-    }
-
-    public String getDescription() {
-      return description;
-    }
-
-    public void setDescription(String description) {
-      this.description = description;
-    }
-
-    public AtomicInteger getNumber() {
-      return number;
-    }
-
-    public void setNumber(AtomicInteger number) {
-      this.number = number;
-    }
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-  }
 }
