+++ src/main/resources/application.properties
@@ -17,7 +17,7 @@ management.metrics.enable.all=false
 # Just showing what is in the contract
 management.metrics.enable.jvm.memory=true
 management.metrics.enable.process.uptime=true
-#management.endpoint.health.enabled=true
+
 endpoints.metrics.sensitive=true
 endpoints.database.sensitive=true
 management.endpoints.web.exposure.include=metrics,database
