+++ src/test/java/urlshortener/web/SystemInformationTests.java
@@ -1,4 +1,4 @@
-package urlshortener.information;
+package urlshortener.web;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
