+++ src/test/resources/application.properties
@@ -16,7 +16,6 @@ management.metrics.enable.all=false
 # Just showing what is in the contract
 management.metrics.enable.jvm.memory=true
 management.metrics.enable.process.uptime=true
-#management.endpoint.health.enabled=true
 endpoints.metrics.sensitive=true
 endpoints.database.sensitive=true
 management.endpoints.web.exposure.include=metrics,database
